+++ lib/providers/trading_provider.dart
@@ -8,7 +8,6 @@ import '../services/trade_csv_export_service.dart';
 import '../services/trade_history_service.dart';
 import '../trading/trading_engine.dart';
 import '../trading/ai_strategy.dart';
-import '../trading/algo_strategy.dart';
 import '../trading/manual_strategy.dart';
 import '../trading/strategy.dart';
 import '../constants/symbols.dart';